Move match card processing to match_card_processor
