+++ scrapers/players/helpers/get_player_information_helper.js
@@ -1,4 +1,5 @@
 const { getText } = require("../../../utils/scrape_utils");
+const { processMatchCard } = require("../../../utils/match_card_processor");
 
 function getMainHeaderData(mainHeader) {
     let playerDetails = {};
@@ -57,44 +58,9 @@ function getPlayerResults(resultsSection) {
     let results = {};
     const matches = resultsSection.children;
 
-    for (let i = 0; i < matches.length - 1; i++) { // Skips the last element since its something different
-        const match = matches[i];
-        let matchDetails = {};
-        const matchID = match.querySelector("a").attributes.href;
+    for (let i = 0; i < matches.length - 1; i++) // Skips the last element since its something different
+        results[i] = processMatchCard(matches[i], true);
 
-        matchDetails.MatchID = matchID.slice(matchID.indexOf('/') + 1, matchID.indexOf('/', 1));
-
-        // Event details
-        matchDetails.EventImage = match.querySelector("div.fc-flex.m-item-thumb").querySelector("img").attributes.src;
-        eventNameContainer = match.querySelector("div.m-item-event").firstElementChild
-        matchDetails.EventName = getText(eventNameContainer);
-        eventNameContainer.remove();
-        matchDetails.SubEvent = getText(match.querySelector("div.m-item-event")).replace(/\s+/g, ' ');
-
-        // Team info
-        matchDetails.TeamAName = getText(match.querySelectorAll("span.m-item-team-name")[0]);
-        matchDetails.TeamBName = getText(match.querySelectorAll("span.m-item-team-name")[1]);
-        matchDetails.TeamAImage = match.querySelector("div.m-item-logo").querySelector("img").attributes.src;
-        matchDetails.TeamBImage = match.querySelector("div.m-item-logo.mod-right").querySelector("img").attributes.src;
-
-        // Score
-        const scoreContainer = match.querySelector("div.m-item-result");
-        matchDetails.TeamAScore = getText(scoreContainer.querySelectorAll("span")[0]);
-        matchDetails.TeamBScore = getText(scoreContainer.querySelectorAll("span")[1]);
-        if (scoreContainer.classList.contains("mod-win"))
-            matchDetails.TeamAWin = true;
-        else
-            matchDetails.TeamAWin = false;
-
-        // Date and time
-        const dateTimeContainer = match.querySelector("div.m-item-date");
-        const dateContainer = dateTimeContainer.querySelector("div");
-        matchDetails.Date = getText(dateContainer);
-        dateContainer.remove();
-        matchDetails.Time = getText(dateTimeContainer).replace(/\s+/g, ' ');
-
-        results[i] = matchDetails;
-    }
     return results;
 }
